Fix Suggestion and Question Forms UI and funtionality for widget
Able to address commit (e1ff887) issues. Both subAccordians google forms 
are displayed and functional. The previous issue was Accordion.Accordion 
prop panels only takes an array of props. Originally, I was passing a 
fucntion that consist of both forms to that array. Now, I pass the 
arrays directly into the array as a value to content attribute.
