+++ app/javascript/src/components/home/HomeBody/index.jsx
@@ -4,7 +4,7 @@ import { Container, Divider, Embed, Transition } from 'semantic-ui-react';
 import DukeTwitterTimeline from './DukeTwitterTimeline';
 import DukeVideo from './DukeVideo';
 import KvilleWeather from './KvilleWeather';
-import Suggestions from './Suggestions';
+import FeedbackForm from './FeedbackForm';
 
 export default class HomeBody extends Component {
   constructor (props) {
@@ -30,7 +30,7 @@ export default class HomeBody extends Component {
           <KvilleWeather />
           {/* <DukeVideo /> */}
           <DukeTwitterTimeline />
-          <Suggestions/>
+          <FeedbackForm />
         </Container>
       </div>
     );
